refactor(ui): move the regex log search out of model.go
The search owns a text-entry mode, its own key semantics, and the blink that
answers a rejected click. That is a distinct surface from the dashboard it
overlays.

Declarations move unchanged.
